Moved database requests to separate file
